requirements.txt
alpaca-py>=0.20.0
python-dotenv>=1.0.0
pandas>=2.0.0
pytz>=2024.1
requests>=2.31.0
anthropic>=0.25.0
